AI/LangChain/SimpleQuestionAnsweringApplication/requirements.txt
# # Modern LangChain dependencies
# langchain==1.0.7
# langchain-openai==1.0.3
# langchain-core==1.0.5

# # OpenAI client library
# openai==2.8.1

# # Environment variable management
# python-dotenv==1.2.1


# # Optional: For better terminal output
# colorama==0.4.6


# Modern LangChain dependencies
langchain==1.0.8
langchain-openai==1.0.3
langchain-core==1.0.7

# OpenAI client library
openai==2.8.1

# Environment variable management
python-dotenv==1.2.1
